move Task model in db module for consistency
